Propose Changes to "italy_cost" Modifier Duration
The current duration for the "italy_cost" modifier is 30 years, This is a VERY long time, as most buildings take a year or two to build. I'm proposing to lower the duration of the modifiers to about 15 years, which gives just enough time for a player to take advantage of the modifier, but not turn Italy into a fully mechanized industrial empire that can stomp competitor nations with ease.